aimedia/requirements.txt
django==2.0.4
PyMySQL==0.8.0
mysqlclient==1.3.12
requests
schedule
pandas
pillow
face_recognition
pytesseract
